+++ ebean-core/src/main/java/io/ebeaninternal/api/SpiQueryBindCapture.java
@@ -1,6 +1,6 @@
 package io.ebeaninternal.api;
 
-import io.ebeaninternal.server.type.bindcapture.BindCapture;
+import io.ebeaninternal.server.bind.capture.BindCapture;
 
 /**
  * Capture query bind values and with those actual database query plans.
