+++ ebean-core/src/main/java/io/ebeaninternal/server/deploy/id/IdBinder.java
@@ -7,7 +7,7 @@ import io.ebeaninternal.server.deploy.BeanProperty;
 import io.ebeaninternal.server.deploy.DbReadContext;
 import io.ebeaninternal.server.deploy.DbSqlContext;
 import io.ebeaninternal.server.query.STreeProperty;
-import io.ebeaninternal.server.type.DataBind;
+import io.ebeaninternal.server.bind.DataBind;
 
 import java.io.DataInput;
 import java.io.DataOutput;
