+++ ebean-core/src/main/java/io/ebeaninternal/server/json/DJsonBeanReader.java
@@ -1,4 +1,4 @@
-package io.ebeaninternal.server.text.json;
+package io.ebeaninternal.server.json;
 
 import com.fasterxml.jackson.core.JsonParser;
 import io.ebean.PersistenceIOException;
