+++ ebean-core/src/main/java/io/ebeaninternal/server/json/DJsonScalar.java
@@ -1,4 +1,4 @@
-package io.ebeaninternal.server.text.json;
+package io.ebeaninternal.server.json;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import io.ebean.core.type.ScalarType;
