+++ ebean-core/src/main/java/io/ebeaninternal/server/json/PathStack.java
@@ -1,4 +1,4 @@
-package io.ebeaninternal.server.text.json;
+package io.ebeaninternal.server.json;
 
 import io.ebeaninternal.server.util.ArrayStack;
