+++ ebean-core/src/main/java/io/ebeaninternal/server/json/ReadJson.java
@@ -1,4 +1,4 @@
-package io.ebeaninternal.server.text.json;
+package io.ebeaninternal.server.json;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
