+++ ebean-core/src/main/java/io/ebeaninternal/server/query/CQueryBindCapture.java
@@ -3,7 +3,7 @@ package io.ebeaninternal.server.query;
 import io.ebeaninternal.api.SpiDbQueryPlan;
 import io.ebeaninternal.api.SpiQueryBindCapture;
 import io.ebeaninternal.api.SpiQueryPlan;
-import io.ebeaninternal.server.type.bindcapture.BindCapture;
+import io.ebeaninternal.server.bind.capture.BindCapture;
 
 import java.util.concurrent.locks.ReentrantLock;
