+++ ebean-core/src/main/java/io/ebeaninternal/server/query/CQueryPlan.java
@@ -11,8 +11,8 @@ import io.ebeaninternal.api.*;
 import io.ebeaninternal.server.core.OrmQueryRequest;
 import io.ebeaninternal.server.core.timezone.DataTimeZone;
 import io.ebeaninternal.server.query.CQueryPlanStats.Snapshot;
-import io.ebeaninternal.server.type.DataBind;
-import io.ebeaninternal.server.type.DataBindCapture;
+import io.ebeaninternal.server.bind.DataBind;
+import io.ebeaninternal.server.bind.DataBindCapture;
 import io.ebeaninternal.server.type.RsetDataReader;
 import io.ebeaninternal.server.util.Md5;
 import io.ebeaninternal.server.util.Str;
