+++ ebean-core/src/main/java/io/ebeaninternal/server/query/QueryPlanLogger.java
@@ -2,7 +2,7 @@ package io.ebeaninternal.server.query;
 
 import io.ebeaninternal.api.SpiDbQueryPlan;
 import io.ebeaninternal.api.SpiQueryPlan;
-import io.ebeaninternal.server.type.bindcapture.BindCapture;
+import io.ebeaninternal.server.bind.capture.BindCapture;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
