+++ ebean-core/src/main/java/io/ebeaninternal/server/querydefn/OrmUpdateProperties.java
@@ -3,7 +3,7 @@ package io.ebeaninternal.server.querydefn;
 import io.ebean.core.type.ScalarType;
 import io.ebeaninternal.server.deploy.DeployParser;
 import io.ebeaninternal.server.persist.Binder;
-import io.ebeaninternal.server.type.DataBind;
+import io.ebeaninternal.server.bind.DataBind;
 
 import java.sql.SQLException;
 import java.util.LinkedHashMap;
