+++ ebean-core/src/main/java/io/ebeanservice/docstore/api/mapping/DocPropertyAdapter.java
@@ -1,6 +1,6 @@
 package io.ebeanservice.docstore.api.mapping;
 
-import io.ebeaninternal.server.text.json.PathStack;
+import io.ebeaninternal.server.json.PathStack;
 
 /**
  * Adapter for DocPropertyVisitor that does not do anything.
