+++ ebean-core/src/test/java/io/ebeaninternal/server/type/JsonTester.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import io.ebean.config.JsonConfig;
 import io.ebean.core.type.ScalarType;
-import io.ebeaninternal.server.text.json.WriteJson;
+import io.ebeaninternal.server.json.WriteJson;
 
 import java.io.IOException;
 import java.io.StringWriter;
