+++ ebean-test/src/test/java/io/ebean/xtest/internal/server/text/json/DJsonScalarTest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import io.ebean.config.DatabaseConfig;
 import io.ebean.platform.h2.H2Platform;
 import io.ebeaninternal.server.core.bootup.BootupClasses;
-import io.ebeaninternal.server.text.json.DJsonScalar;
+import io.ebeaninternal.server.json.DJsonScalar;
 import io.ebeaninternal.server.type.DefaultTypeManager;
 import org.junit.jupiter.api.Test;
